Conflicts:
	src/main/frontend/src/app/configuration/employee-config/employee-config.component.html
	src/main/java/com/apmosys/employeeportal/dto/EmployeeDTO.java
	src/main/java/com/apmosys/employeeportal/service/EmployeeLeaveService.java
	src/main/java/com/apmosys/employeeportal/service/EmployeeService.java

+++ src/main/java/com/apmosys/employeeportal/service/EmployeeLeaveService.java
@@ -308,7 +308,7 @@ public class EmployeeLeaveService {
 				    List<Timesheet> empTimeSheet = timesheetsRepository.findTimesheetOnLeaveDate(leaveDTO.getEmpId(),leaveDTO.getFromDate(),leaveDTO.getToDate());
 				    if(!empTimeSheet.isEmpty()) {
 				    empTimeSheet.forEach((timesheet)->{
-				    	
+				     	
 				    	List<TimesheetActivityMap> timesheetactivities = timesheetActivityRepository.getTimesheetActivityByTimesheetId(timesheet.getTimesheetId());
 				    	
 				    	timesheetactivities.forEach((timesheetactivity)->{
@@ -317,7 +317,9 @@ public class EmployeeLeaveService {
 
 				    	});
 				    	
-				    	timesheetsRepository.deleteById(timesheet.getTimesheetId());
+				    	if(!timesheet.getDayType().equals("Public Holiday") && !timesheet.getDayType().equals("Week Off")) {
+				    		timesheetsRepository.deleteById(timesheet.getTimesheetId());				    		
+				    	}
 				    });
 				    }
 				    //after timesheet deletion
@@ -1180,54 +1182,52 @@ public class EmployeeLeaveService {
 					leaveDTO.getEmployeeLeaveList().forEach((dto) -> {
 
 						if (leave.getLeaveTypeMasterId() == dto.getLeaveTypeMasterId()) {
-
-							if (leave.getBalance() == 0) {
-								// set leaves to newly created employee when his/her bucket is 0.0 for all leave
-								// types
-								LeaveBalanceLog log = new LeaveBalanceLog();
-								log.setBalance(dto.getBalance());
-								log.setEmpId(leaveDTO.getEmpId());
-								log.setLeaveTypeMasterId(dto.getLeaveTypeMasterId());
-								log.setMessage(
-										LeaveLogMessage.adminAddLeave.replace("0.0", dto.getBalance().toString()));
-								log.setUpdateBalanceBy("+" + dto.getBalance());
-								leaveBalanceLogRepository.save(log);
-
-								leave.setBalance(dto.getBalance());
-							} else {
-
-								if (dto.getBalance().equals(leave.getBalance())) {
-									// No change in balance leave
-									leave.setBalance(dto.getBalance());
-								} else {
-
+							if(dto.getBalance() != 0) {
+								if (leave.getBalance() == 0) {
+									// set leaves to newly created employee when his/her bucket is 0.0 for all leave
+									// types
 									LeaveBalanceLog log = new LeaveBalanceLog();
 									log.setBalance(dto.getBalance());
 									log.setEmpId(leaveDTO.getEmpId());
 									log.setLeaveTypeMasterId(dto.getLeaveTypeMasterId());
-
-									if (dto.getBalance() > leave.getBalance()) {
-										// leave added to bucket balance
-										Float change = dto.getBalance() - leave.getBalance();
-										log.setMessage(LeaveLogMessage.adminAddLeave.replace("0.0", change.toString()));
-										log.setUpdateBalanceBy("+" + change);
-
-									} else if (dto.getBalance() < leave.getBalance()) {
-										// leave deducted from bucket balance
-										Float change = dto.getBalance() - leave.getBalance();
-										log.setMessage(
-												LeaveLogMessage.adminDeductLeave.replace("0.0", (change * -1) + ""));
-										log.setUpdateBalanceBy(change.toString());
-									}
-									leave.setBalance(dto.getBalance());
+									log.setMessage(
+											LeaveLogMessage.adminAddLeave.replace("0.0", dto.getBalance().toString()));
+									log.setUpdateBalanceBy("+" + dto.getBalance());
 									leaveBalanceLogRepository.save(log);
+
+									leave.setBalance(dto.getBalance());
+								} else {
+
+									if (dto.getBalance().equals(leave.getBalance())) {
+										// No change in balance leave
+										leave.setBalance(dto.getBalance());
+									} else {
+
+										LeaveBalanceLog log = new LeaveBalanceLog();
+										log.setBalance(dto.getBalance());
+										log.setEmpId(leaveDTO.getEmpId());
+										log.setLeaveTypeMasterId(dto.getLeaveTypeMasterId());
+
+										if (dto.getBalance() > leave.getBalance()) {
+											// leave added to bucket balance
+											Float change = dto.getBalance() - leave.getBalance();
+											log.setMessage(LeaveLogMessage.adminAddLeave.replace("0.0", change.toString()));
+											log.setUpdateBalanceBy("+" + change);
+
+										} else if (dto.getBalance() < leave.getBalance()) {
+											// leave deducted from bucket balance
+											Float change = dto.getBalance() - leave.getBalance();
+											log.setMessage(
+													LeaveLogMessage.adminDeductLeave.replace("0.0", (change * -1) + ""));
+											log.setUpdateBalanceBy(change.toString());
+										}
+										leave.setBalance(dto.getBalance());
+										leaveBalanceLogRepository.save(log);
+									}
 								}
 							}
-
 						}
-
 					});
-
 				});
 			}
 
@@ -2918,57 +2918,6 @@ public ServiceResponse getEmployeeLeaveApplicationwithHolidays(LeaveDTO leaveDTO
 		return response;
 	}
 
-	public ServiceResponse addTimesheetForHolidays(LeaveDTO leaveDTO) {
-		ServiceResponse response = new ServiceResponse();
-		try {
-			
-			LocalDate holidayDate = LocalDate.parse("2023-03-22");
-			
-			List<Employee> allEmployee = employeeRepository.findAll();
-			
-			if(!allEmployee.isEmpty()) {
-				allEmployee.forEach((emp) -> {
-					
-					if(!emp.getEmploymentstatus().equals("InActive")) {
-						Timesheet timesheetObj = timesheetsRepository.findByEmpIdAndDate(emp.getEmpId(), holidayDate);
-						
-						if(timesheetObj == null) {
-							Timesheet newTimesheet = new Timesheet();
-							
-							newTimesheet.getCommonProperty().setCreatedBy(emp.getEmpId());
-							newTimesheet.setDate(holidayDate);
-							newTimesheet.setDayType("Holiday");
-							newTimesheet.setDescription("Public Holiday");
-							newTimesheet.setEmpId(emp.getEmpId());
-							newTimesheet.setStatus("Approved");
-							
-							Timesheet dbResponse = timesheetsRepository.save(newTimesheet);
-							
-							if(dbResponse != null) {
-								response.setServiceStatus(ServiceResponse.STATUS_SUCCESS);
-								response.setServiceResponse("timesheet Added successfully");
-							}else {
-								response.setServiceStatus(ServiceResponse.STATUS_FAIL);
-								response.setServiceResponse("Unable to add Timesheet");
-							}
-						}
-					}
-				});
-			}else {
-				response.setServiceStatus(ServiceResponse.STATUS_FAIL);
-				response.setServiceResponse("Employee List is empty");
-			}
-			
-			
-		}catch(Exception e) {
-			e.printStackTrace();
-			response.setServiceStatus(ServiceResponse.SOMETHING_WENT_WRONG);
-			response.setServiceResponse("Something Went Wrong.");
-			response.setServiceError(e.getMessage());
-		}
-		return response;
-	}
-
 	public ServiceResponse pendingForApprovalReconsilation() {
 		ServiceResponse response = new ServiceResponse();
 		try {
@@ -3025,6 +2974,84 @@ public ServiceResponse getEmployeeLeaveApplicationwithHolidays(LeaveDTO leaveDTO
 			
 			response.setServiceStatus(ServiceResponse.STATUS_SUCCESS);
 			response.setServiceResponse(dtoList);
+			
+		}catch(Exception e) {
+			e.printStackTrace();
+			response.setServiceStatus(ServiceResponse.SOMETHING_WENT_WRONG);
+			response.setServiceResponse("Something Went Wrong.");
+			response.setServiceError(e.getMessage());
+		}
+		return response;
+	}
+
+	public ServiceResponse reconsileCasualBalance() {
+		ServiceResponse response = new ServiceResponse();
+		try {
+			
+			List<Employee> allEmployee = employeeRepository.findAll();
+			
+			List<Employee> json = new ArrayList<>();
+			
+			if(!allEmployee.isEmpty()) {
+				allEmployee.forEach((object) -> {
+					List<LeaveBalanceLog> balanceLogObj = leaveBalanceLogRepository
+							.findLogToReconsile(object.getEmpId(), (short) 2);
+					
+					Float updatedBalance = (float) 0;
+					
+					for(LeaveBalanceLog logObj : balanceLogObj) {
+						String message = logObj.getUpdateBalanceBy();
+						float number = Float.parseFloat(message.replaceAll("[^\\d.]", ""));
+						
+						System.out.println(number + " number \n\n\n");
+						
+						updatedBalance = updatedBalance + number;
+					}
+					
+					EmployeeLeavesMap empLeaveMap = employeeLeavesMapRepository
+							.findByEmpIdAndLeaveTypeMasterId(object.getEmpId(), (short) 2);
+					
+					if(empLeaveMap != null) {
+						empLeaveMap.setBalance(updatedBalance);
+						EmployeeLeavesMap dbResponse = employeeLeavesMapRepository.save(empLeaveMap);		
+						
+						if(dbResponse != null) {
+							LeaveBalanceLog log = new LeaveBalanceLog();
+
+							log.setBalance(updatedBalance);
+							log.setEmpId(object.getEmpId());
+							log.setLeaveTypeMasterId((short) 2);
+							log.setMessage(LeaveLogMessage.autoAddLeave.replace("0.0",
+									Float.toString(updatedBalance)));
+							log.setUpdateBalanceBy("-" + updatedBalance);
+
+							LeaveBalanceLog logResponse = leaveBalanceLogRepository.save(log);
+							
+							if(logResponse != null) {
+								
+//								Employee newObj = new Employee();
+//								newObj.setEmpId(object.getEmpId());
+//								newObj.setEmployeementId(object.getEmployeementId());
+//								newObj.setName(object.getName());
+//								newObj.setPreviousLog(balanceLogObj.get(0).getMessage());
+//								newObj.setNewLog(log.getMessage());
+//								
+//								json.add(newObj);
+								
+								response.setServiceStatus(ServiceResponse.STATUS_SUCCESS);
+								response.setServiceResponse("Casual Leave recosiled Successfully.");
+							}else {
+								response.setServiceStatus(ServiceResponse.STATUS_FAIL);
+								response.setServiceResponse("Unable to reconsile Casual Leave.");
+							}
+						}
+					}
+					
+				});
+			}
+			
+			response.setServiceStatus(ServiceResponse.STATUS_SUCCESS);
+			response.setServiceResponse1(json);
 			
 		}catch(Exception e) {
 			e.printStackTrace();
